update examdetail, udpate backend for create new exam

+++ Examination_DAL/Models/ExamDetail.cs
@@ -38,7 +38,7 @@ namespace Examination_DAL.Models
         public int TotalQuestion { get; set; }
 
         [Required]
-        public int MaxiumMark { get; set; }
+        public double MaxiumMark { get; set; }
 
         [AllowNull]
         public int? ReTestNumber { get; set; }
